# clients/typescript/openapi.json
{
  "components": {
    "schemas": {
      "Action": {
        "additionalProperties": true,
        "description": "laya extra: how strongly the model would act on this answer.",
        "properties": {
          "act_probability": {
            "title": "Act Probability",
            "type": "number"
          }
        },
        "required": [
          "act_probability"
        ],
        "title": "Action",
        "type": "object"
      },
      "ChoiceAnswer": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "choice": {
            "description": "The selected option key",
            "title": "Choice",
            "type": "string"
          },
          "confidence": {
            "description": "How peaked the distribution is, 0-1",
            "title": "Confidence",
            "type": "number"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Distribution over your options",
            "title": "Probabilities",
            "type": "object"
          },
          "type": {
            "const": "choice",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "choice",
          "probabilities",
          "confidence"
        ],
        "title": "ChoiceAnswer",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved: \"cuda\" or \"cpu\"",
            "title": "Device"
          },
          "device_requested": {
            "description": "\"auto\", \"cpu\" or \"cuda\"",
            "title": "Device Requested",
            "type": "string"
          },
          "gpu": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "models",
          "device_requested"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "NoulAnswer": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "noul": {
            "description": "Probability the statement is true, 0-1",
            "title": "Noul",
            "type": "number"
          },
          "type": {
            "const": "noul",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "noul"
        ],
        "title": "NoulAnswer",
        "type": "object"
      },
      "Question": {
        "additionalProperties": true,
        "description": "One typed question. Shape is identical to Jev's.",
        "properties": {
          "criteria": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "choice: {option: meaning}. score: ordered levels, low to high. noul: optional clarification of yes and no.",
            "title": "Criteria"
          },
          "instructions": {
            "description": "The judgement to make; a string, object or list",
            "title": "Instructions"
          },
          "type": {
            "enum": [
              "choice",
              "score",
              "noul"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "instructions"
        ],
        "title": "Question",
        "type": "object"
      },
      "Routing": {
        "additionalProperties": true,
        "description": "laya extra: which checkpoint served the request, and why.",
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "Routing",
        "type": "object"
      },
      "ScoreAnswer": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Action"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "legend": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Your levels, keyed by index",
            "title": "Legend",
            "type": "object"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Distribution over levels",
            "title": "Probabilities",
            "type": "object"
          },
          "score": {
            "description": "Position along your levels; may fall between two",
            "title": "Score",
            "type": "number"
          },
          "type": {
            "const": "score",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "score",
          "legend",
          "probabilities",
          "confidence"
        ],
        "title": "ScoreAnswer",
        "type": "object"
      },
      "SystemOneRequest": {
        "additionalProperties": true,
        "examples": [
          {
            "questions": {
              "churn_risk": {
                "instructions": "The customer threatens to leave for a competitor",
                "type": "noul"
              },
              "department": {
                "criteria": {
                  "billing": "Payment or subscription issues",
                  "sales": "Pricing or account questions",
                  "technical": "Bugs or integration problems"
                },
                "instructions": "Which team should handle this?",
                "type": "choice"
              },
              "frustration": {
                "criteria": [
                  "Calm, just stating facts",
                  "Frustrated but civil",
                  "Very angry, strong language"
                ],
                "instructions": "How frustrated the customer appears",
                "type": "score"
              }
            },
            "state": "Hi, we were billed twice for March and I've had no reply in 3 days. If this isn't fixed we'll move to a competitor."
          }
        ],
        "properties": {
          "lang": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Routing hint, e.g. 'en'; otherwise detected",
            "title": "Lang"
          },
          "model": {
            "anyOf": [
              {
                "enum": [
                  "english",
                  "multilingual",
                  "typed-decisions"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Force a checkpoint; omit to auto-route",
            "title": "Model"
          },
          "questions": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Question"
            },
            "description": "Typed questions keyed by the id you want back in `answers`",
            "title": "Questions",
            "type": "object"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "description": "Text, JSON object, or conversation turns",
            "title": "State"
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Routing hint, e.g. 'typed-decisions'",
            "title": "Task"
          }
        },
        "required": [
          "state",
          "questions"
        ],
        "title": "SystemOneRequest",
        "type": "object"
      },
      "SystemOneResponse": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "discriminator": {
                "mapping": {
                  "choice": "#/components/schemas/ChoiceAnswer",
                  "noul": "#/components/schemas/NoulAnswer",
                  "score": "#/components/schemas/ScoreAnswer"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ChoiceAnswer"
                },
                {
                  "$ref": "#/components/schemas/ScoreAnswer"
                },
                {
                  "$ref": "#/components/schemas/NoulAnswer"
                }
              ]
            },
            "title": "Answers",
            "type": "object"
          },
          "model": {
            "description": "The checkpoint that answered",
            "title": "Model",
            "type": "string"
          },
          "routing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Routing"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "model",
          "answers",
          "usage"
        ],
        "title": "SystemOneResponse",
        "type": "object"
      },
      "Usage": {
        "properties": {
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          }
        },
        "required": [
          "input_tokens",
          "output_tokens"
        ],
        "title": "Usage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "laya-serve",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/route": {
      "post": {
        "description": "Routing decision only \u2014 no forward pass, sub-millisecond.",
        "operationId": "route_v1_route_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemOneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Routing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route"
      }
    },
    "/v1/systemone": {
      "post": {
        "operationId": "system_one_v1_systemone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemOneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "System One"
      }
    }
  }
}